[Minor] Select types update
